+++ test/OpenTelemetry.Exporter.Zipkin.Tests/ZipkinExporterTests.cs
@@ -109,7 +109,7 @@ namespace OpenTelemetry.Exporter.Zipkin.Tests
                 Endpoint = new Uri($"http://{this.testServerHost}:{this.testServerPort}/api/v2/spans?requestId={requestId}"),
             };
             var zipkinExporter = new ZipkinExporter(exporterOptions);
-            var exportActivityProcessor = new BatchExportProcessor<Activity>(zipkinExporter);
+            var exportActivityProcessor = new BatchActivityExportProcessor(zipkinExporter);
 
             var openTelemetrySdk = Sdk.CreateTracerProviderBuilder()
                 .AddSource(ActivitySourceName)
@@ -196,7 +196,7 @@ namespace OpenTelemetry.Exporter.Zipkin.Tests
                 activity.SetTag(ZipkinActivityConversionExtensions.ZipkinErrorFlagTagName, "This should be removed.");
             }
 
-            var processor = new SimpleExportProcessor<Activity>(exporter);
+            var processor = new SimpleActivityExportProcessor(exporter);
 
             processor.OnEnd(activity);